test(e2e): extract shared helpers and group specs into thematic directories
