+++ .gitignore
@@ -1,5 +1,7 @@
 # Logs
 logs
+# The bare 'logs' pattern above is for runtime log output, not the logs E2E suite.
+!tests/e2e/logs/
 *.log
 npm-debug.log*
 yarn-debug.log*
