+++ cspell.config.json
@@ -17,8 +17,10 @@
     "aggregatable",
     "aheads",
     "alertbox",
-    "ASOF",
     "apikey",
+    "ASOF",
+    "behaviour",
+    "Brien",
     "buildinfo",
     "chcol",
     "CHSQL",
@@ -33,15 +35,16 @@
     "dataframe",
     "dataproxy",
     "DataSource",
-    "datasourcese",
     "DataSources",
-    "Drilldown",
+    "datasourcese",
+    "datepart",
     "DateTime",
     "dbug",
     "dedup",
-    "downsamples",
     "downsampled",
+    "downsamples",
     "downsampling",
+    "Drilldown",
     "elazarl",
     "emerg",
     "endregion",
@@ -56,41 +59,43 @@
     "groupable",
     "healthcheck",
     "ILIKE",
+    "initialise",
     "instancemgmt",
     "jsonencode",
     "listbox",
     "lowcardinality",
     "lttb",
     "macropro",
-    "datepart",
     "magefile",
     "misordered",
-    "multiquery",
     "mgbench",
     "Milli",
     "millis",
     "moby",
+    "multiquery",
     "mydb",
     "networkidle",
     "nofile",
-    "noreferrer",
     "nolint",
-    "operationname",
     "Nonproxy",
+    "noreferrer",
+    "OFREP",
     "openfeature",
     "oper",
-    "OFREP",
+    "operationname",
+    "optimisation",
+    "optimised",
     "otel",
     "otelcol",
-    "pageerror",
-    "semconv",
     "OUTFILE",
+    "pageerror",
     "paulmach",
     "pgsql",
     "picklist",
     "picomatch",
     "PREWHERE",
     "proxying",
+    "rederives",
     "regexes",
     "retarget",
     "retargets",
@@ -98,18 +103,21 @@
     "schemads",
     "sdkconfig",
     "sdkproxy",
+    "semconv",
     "servicename",
     "shopspring",
     "singlequote",
     "slvrtrn",
     "sparkline",
+    "spinbutton",
     "sqlds",
+    "sqlstring",
     "sqlutil",
     "stagename",
     "stretchr",
     "subkey",
-    "subquery",
     "subqueries",
+    "subquery",
     "subresource",
     "sugg",
     "supress",
@@ -131,11 +139,14 @@
     "typecheck",
     "Ulimits",
     "unconfigured",
+    "undercounts",
     "uuidv",
     "uvcf",
-    "vectorator",
-    "WorkDir",
     "varname",
+    "vectorator",
+    "virtualised",
+    "visualisation",
+    "WorkDir",
     "zizmor"
   ]
 }
