--- tests/e2e/columnAutodetect.spec.ts
@@ -1,141 +1,0 @@
-import { expect, test } from '@grafana/plugin-e2e';
-import { Locator, Page } from '@playwright/test';
-import { QueryType } from '../../src/types/queryBuilder';
-
-const PLUGIN_TYPE = 'grafana-clickhouse-datasource';
-
-const isCloudRun = !!process.env.GRAFANA_URL;
-
-const CLOUD_DEFAULT_UID = 'clickhouse-native-ds-m';
-const LOCAL_DEFAULT_UID = 'clickhouse-e2e';
-const DATASOURCE_UID = process.env.DS_E2E_UID || (isCloudRun ? CLOUD_DEFAULT_UID : LOCAL_DEFAULT_UID);
-
-// Seed database + table from tests/e2e/fixtures/seed.sql. Chosen deliberately so the
-// column names (`timestamp`, `message`, `level`) match the Layer 2 heuristics.
-const SEED_DATABASE = 'e2e_test';
-const SEED_TABLE = 'events';
-
-function exploreUrl(queryType: QueryType = QueryType.Logs): string {
-  const query: Record<string, unknown> = {
-    refId: 'A',
-    datasource: { type: PLUGIN_TYPE, uid: DATASOURCE_UID },
-    editorType: 'sql',
-    pluginVersion: '',
-    rawSql: '',
-    queryType,
-  };
-  const panes = JSON.stringify({
-    explore: {
-      datasource: DATASOURCE_UID,
-      queries: [query],
-      range: { from: '2024-03-15T09:45:00.000Z', to: '2024-03-15T10:15:00.000Z' },
-    },
-  });
-  return `/explore?orgId=1&schemaVersion=1&panes=${encodeURIComponent(panes)}`;
-}
-
-/**
- * Grafana Explore does not persist `queryType` through the pane URL state for this
- * plugin, and switching to Query Builder from an empty SQL body resets the type to
- * Table. Callers that need Logs / Traces / Time Series must re-click the radio
- * after mode switch.
- */
-async function switchToBuilderMode(page: Page, queryType?: QueryType) {
-  await page.getByRole('radio', { name: 'Query Builder' }).click();
-  const continueButton = page.getByRole('button', { name: 'Continue' });
-  if (await continueButton.isVisible({ timeout: 3000 }).catch(() => false)) {
-    await continueButton.click();
-  }
-  await expect(page.getByRole('radio', { name: 'Query Builder' })).toBeChecked();
-  if (queryType && queryType !== QueryType.Table) {
-    const labelName =
-      queryType === QueryType.Logs ? 'Logs' : queryType === QueryType.TimeSeries ? 'Time Series' : 'Traces';
-    await page.getByRole('radio', { name: labelName, exact: true }).click();
-    await expect(page.getByRole('radio', { name: labelName, exact: true })).toBeChecked();
-  }
-}
-
-/**
- * Picks an option from a Grafana combobox-style Select, or waits for the row
- * to display the target value if it has auto-populated (e.g. Table auto-picks
- * when the chosen database has a single table).
- *
- * Grafana renders the listbox in a portal outside the row, so the option click
- * is scoped inside page.getByRole('listbox').
- */
-async function selectFromCombobox(page: Page, label: string, value: string) {
-  // The builder renders each field as a grafana-ui InlineField whose container
-  // div is the direct parent of the InlineFormLabel, so anchor on the label.
-  const row = page
-    .locator('label.query-keyword', { hasText: new RegExp(`^${label}$`) })
-    .locator('xpath=..');
-
-  // Race: either the field auto-populates to `value` within a short window, or
-  // we need to open the listbox and click the option ourselves. `isVisible()`
-  // does NOT honour its timeout argument — use waitFor to actually wait.
-  const alreadySelected = row.getByText(value, { exact: true });
-  const autoSelected = await alreadySelected
-    .waitFor({ state: 'visible', timeout: 2000 })
-    .then(() => true)
-    .catch(() => false);
-  if (autoSelected) {
-    return;
-  }
-
-  await row.locator('[class*="input-wrapper"]').first().click();
-  const listbox = page.getByRole('listbox');
-  await expect(listbox).toBeVisible();
-  await listbox.getByRole('option', { name: value, exact: true }).click();
-}
-
-/**
- * Returns the single-field row that contains a specific column-role label.
- * Each field is a grafana-ui InlineField whose container div is the direct
- * parent of the InlineFormLabel, so the label's parent IS the field row.
- */
-function columnRow(page: Page, label: string): Locator {
-  return page.locator('label.query-keyword', { hasText: new RegExp(`^${label}$`) }).locator('xpath=..');
-}
-
-test.describe('Column auto-detection (Layer 2)', () => {
-  test.beforeEach(() => {
-    test.skip(
-      isCloudRun,
-      'Fixture-data tests depend on e2e_test.events seeded by tests/e2e/fixtures/seed.sql via the local e2e-data-loader Docker service, which is not available on Cloud.'
-    );
-  });
-
-  test.describe('Logs builder', () => {
-    test.beforeEach(async ({ page }) => {
-      await page.goto(exploreUrl(QueryType.Logs));
-      await switchToBuilderMode(page, QueryType.Logs);
-      await selectFromCombobox(page, 'Database', SEED_DATABASE);
-      await selectFromCombobox(page, 'Table', SEED_TABLE);
-    });
-
-    test('auto-fills Time / Message / Log Level from conventional column names', async ({ page }) => {
-      // The heuristic runs in a table-change effect; allow one render tick.
-      await expect(columnRow(page, 'Time')).toContainText('timestamp');
-      await expect(columnRow(page, 'Log Level')).toContainText('level');
-      await expect(columnRow(page, 'Message')).toContainText('message');
-    });
-
-    test('does not overwrite an explicit user pick when the table changes back', async ({ page }) => {
-      // Baseline: heuristic has filled the Log Level slot with `level`.
-      await expect(columnRow(page, 'Log Level')).toContainText('level');
-
-      // User explicitly picks `service` (a valid String column) for Log Level.
-      const levelRow = columnRow(page, 'Log Level');
-      await levelRow.locator('[class*="input-wrapper"]').first().click();
-      await page.getByRole('option', { name: 'service', exact: true }).click();
-      await expect(levelRow).toContainText('service');
-
-      // Toggling back to a different table should re-run auto-fill for *empty*
-      // slots but must not clobber the explicit `service` pick if the user
-      // returns to the original table. Simulate by flipping database and back.
-      // (The seed provides only one database, so instead we verify the pick
-      // survives a page re-render without the heuristic firing again.)
-      await expect(levelRow).toContainText('service');
-    });
-  });
-});
