+++ tests/e2e/traces/traceQueryNoCrash.spec.ts
@@ -1,5 +1,7 @@
-import { expect, test, ExplorePage } from '@grafana/plugin-e2e';
-import { Page } from '@playwright/test';
+import { expect, test } from '@grafana/plugin-e2e';
+import type { Page } from '@playwright/test';
+import { FIXTURE_FROM_ISO, FIXTURE_TO_ISO, skipFixtureTestsOnCloud } from '../helpers/env';
+import { exploreUrl } from '../helpers/explore';
 
 // Regression guard for the Traces query-builder crash on older Grafana
 // (reproduced on the >=11.6 floor, worked on 13.x).
@@ -20,19 +22,6 @@ import { Page } from '@playwright/test';
 // depth" render loops, etc.) are caught here too — not just this exact bug.
 // Its value is the cross-version matrix: the reusable plugin CI runs Playwright
 // against the grafanaDependency floor, where this class of crash manifests.
-
-const PLUGIN_TYPE = 'grafana-clickhouse-datasource';
-
-const isCloudRun = !!process.env.GRAFANA_URL;
-
-const CLOUD_DEFAULT_UID = 'clickhouse-native-ds-m';
-const LOCAL_DEFAULT_UID = 'clickhouse-e2e';
-const DATASOURCE_UID = process.env.DS_E2E_UID || (isCloudRun ? CLOUD_DEFAULT_UID : LOCAL_DEFAULT_UID);
-
-// trace_spans (tests/e2e/fixtures/trace_spans.sql) seeds spans for e2e-trace-a
-// and e2e-trace-b between 2024-03-15 10:00:00 and 10:00:10 UTC.
-const FIXTURE_FROM_ISO = '2024-03-15T09:45:00.000Z';
-const FIXTURE_TO_ISO = '2024-03-15T10:15:00.000Z';
 
 // Fatal client-side error signatures. Matching any of these means a render or
 // transform blew up rather than failing gracefully.
@@ -65,77 +54,61 @@ function collectClientErrors(page: Page): { getErrors: () => string[] } {
   return { getErrors: () => errors };
 }
 
-/**
- * Builds an Explore URL for a Traces query-builder query against trace_spans.
- * A builder query (queryType=traces) is used deliberately: it always attaches
- * the View trace / View logs data links — the exact path that crashed — without
- * needing datasource-level trace defaults.
- */
-function traceExploreUrl(): string {
-  const builderOptions = {
-    database: 'e2e_test',
-    table: 'trace_spans',
-    queryType: 'traces',
-    mode: 'list',
-    columns: [
-      { name: 'TraceId', type: 'String', hint: 'trace_id' },
-      { name: 'SpanId', type: 'String', hint: 'trace_span_id' },
-      { name: 'ParentSpanId', type: 'String', hint: 'trace_parent_span_id' },
-      { name: 'ServiceName', type: 'LowCardinality(String)', hint: 'trace_service_name' },
-      { name: 'SpanName', type: 'LowCardinality(String)', hint: 'trace_operation_name' },
-      { name: 'Timestamp', type: 'DateTime64(9)', hint: 'time' },
-      { name: 'Duration', type: 'Int64', hint: 'trace_duration_time' },
-    ],
-    meta: {},
-    limit: 1000,
-    filters: [],
-    orderBy: [],
-  };
+// Builder options for a Traces query against trace_spans (fixture
+// tests/e2e/fixtures/trace_spans.sql seeds spans for e2e-trace-a and
+// e2e-trace-b between 2024-03-15 10:00:00 and 10:00:10 UTC). A builder query
+// (queryType=traces) is used deliberately: it always attaches the View trace /
+// View logs data links — the exact path that crashed — without needing
+// datasource-level trace defaults.
+const TRACE_BUILDER_OPTIONS = {
+  database: 'e2e_test',
+  table: 'trace_spans',
+  queryType: 'traces',
+  mode: 'list',
+  columns: [
+    { name: 'TraceId', type: 'String', hint: 'trace_id' },
+    { name: 'SpanId', type: 'String', hint: 'trace_span_id' },
+    { name: 'ParentSpanId', type: 'String', hint: 'trace_parent_span_id' },
+    { name: 'ServiceName', type: 'LowCardinality(String)', hint: 'trace_service_name' },
+    { name: 'SpanName', type: 'LowCardinality(String)', hint: 'trace_operation_name' },
+    { name: 'Timestamp', type: 'DateTime64(9)', hint: 'time' },
+    { name: 'Duration', type: 'Int64', hint: 'trace_duration_time' },
+  ],
+  meta: {},
+  limit: 1000,
+  filters: [],
+  orderBy: [],
+};
 
-  const query = {
-    refId: 'A',
-    datasource: { type: PLUGIN_TYPE, uid: DATASOURCE_UID },
-    editorType: 'builder',
-    pluginVersion: '',
-    // The builder normally generates this; supply it so the query runs
-    // deterministically on load without depending on editor re-generation.
-    rawSql:
-      'SELECT "TraceId" as traceID, "ServiceName" as serviceName, "SpanName" as operationName, ' +
-      '"Timestamp" as startTime, "Duration" as duration FROM "e2e_test"."trace_spans" ' +
-      'WHERE ( "Timestamp" >= $__fromTime AND "Timestamp" <= $__toTime ) ORDER BY "Timestamp" DESC LIMIT 1000',
-    builderOptions,
-    format: 1,
-  };
-
-  const panes = JSON.stringify({
-    explore: {
-      datasource: DATASOURCE_UID,
-      queries: [query],
-      range: { from: FIXTURE_FROM_ISO, to: FIXTURE_TO_ISO },
-    },
-  });
-  return `/explore?orgId=1&schemaVersion=1&panes=${encodeURIComponent(panes)}`;
-}
+// The builder normally generates this; supply it so the query runs
+// deterministically on load without depending on editor re-generation.
+const TRACE_RAW_SQL =
+  'SELECT "TraceId" as traceID, "ServiceName" as serviceName, "SpanName" as operationName, ' +
+  '"Timestamp" as startTime, "Duration" as duration FROM "e2e_test"."trace_spans" ' +
+  'WHERE ( "Timestamp" >= $__fromTime AND "Timestamp" <= $__toTime ) ORDER BY "Timestamp" DESC LIMIT 1000';
 
 test.describe('Traces query renders without crashing', () => {
   test.beforeEach(() => {
-    test.skip(
-      isCloudRun,
-      'Fixture-data tests depend on the local trace_spans seed (tests/e2e/fixtures/trace_spans.sql) loaded via the e2e-data-loader Docker service, which is not available on Cloud.'
-    );
+    skipFixtureTestsOnCloud('trace_spans.sql');
   });
+
+  test.describe.configure({ mode: 'serial' });
 
   test('a Traces query with rendered trace-detail links throws no uncaught/fatal client errors', async ({
     page,
     explorePage,
-  }: {
-    page: Page;
-    explorePage: ExplorePage;
   }) => {
     const { getErrors } = collectClientErrors(page);
 
     const responsePromise = explorePage.waitForQueryDataResponse();
-    await page.goto(traceExploreUrl());
+    await page.goto(
+      exploreUrl({
+        from: FIXTURE_FROM_ISO,
+        to: FIXTURE_TO_ISO,
+        rawSql: TRACE_RAW_SQL,
+        builderOptions: TRACE_BUILDER_OPTIONS,
+      })
+    );
     await responsePromise;
 
     // Let the results panel render and any deferred link scan run. The crash
